"Refactor population statistics service and tests.
Decoupled population statistics service from the main ItemSummary projection, splitting it into three different projections: quality, type, and slot. This change makes each section more self-contained and maintainable. Test coverage has also been expanded to ensure the correct statistics are calculated. Updated error messages to provide more specific feedback."

+++ src/main/java/com/thoroldvix/economatic/summary/item/ItemSummaryResponse.java
@@ -15,7 +15,7 @@ public record ItemSummaryResponse(
 
             Type type,
 
-            int total
+            long total
     ) {
 
     }
